\name{surv2sample-internal}
\alias{multiplot}
\alias{print.neyman.test}
\alias{summary.neyman.test}
\alias{print.surv2.ks}
\alias{print.surv2.logrank}
\alias{print.surv2.neyman}
\alias{print.Survcomp}
\alias{print.cif}
\alias{print.cif2.int}
\alias{print.cif2.ks}
\alias{print.cif2.logrank}
\alias{print.cif2.neyman}
\alias{print.proprate2.fit}
\alias{print.proprate2.ks}
\alias{print.proprate2.gs}
\alias{print.proprate2.neyman}
\title{Internal Functions for the Package surv2sample}
\description{
Internal functions for the package \pkg{surv2sample}.
Usually not to be directly called by users.
}
\author{David Kraus (\url{http://www.davidkraus.net/})}
\keyword{internal}
